s/resiliency/resilience/, round 1
From a grammatical perspective, "resilience" and "resiliency" are generally accepted as the same thing, but notably "usage authorities consider it a needless variant. ... [R]esilience is much more common and has been for at least 100 years." [0]

But potentially more compelling: the field of study most likely to inform this SIG is Resilience (capital R) Engineering; so we'd do well to use that word of they -ey variant.

[0] https://writingexplained.org/resilience-vs-resiliency-difference